+++ example/lib/main.dart
@@ -21,7 +21,7 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, required this.title}) : super(key: key);
   final String title;
 
   @override
@@ -52,7 +52,7 @@ class _MyHomePageState extends State<MyHomePage> {
 }
 
 class NextPage extends StatefulWidget {
-  NextPage({Key key}) : super(key: key);
+  NextPage({Key? key}) : super(key: key);
 
   @override
   NextPageState createState() => NextPageState();
@@ -82,21 +82,20 @@ class NextPageState extends State<NextPage> {
       body: MultiPull(
         circleMoveDuration: Duration(milliseconds: 400),
         circleMoveCurve: Curves.easeInOut,
-        actionWidgets: [
-          ActionWidget(
+        pullIndicators: [
+          PullIndicator(
             icon: Icon(Icons.arrow_back_ios_outlined),
             label: Text("back"),
-            action: () => Navigator.pop(context),
+            onPull: () => Navigator.pop(context),
           ),
-          ActionWidget(
+          PullIndicator(
             icon: Icon(Icons.refresh_rounded),
-            // label: Text("reload"),
-            onRefresh: () async => await Future.delayed(Duration(seconds: 2)),
+            onPull: () async => await Future.delayed(Duration(seconds: 2)),
           ),
-          ActionWidget(
+          PullIndicator(
             icon: Icon(Icons.backspace_outlined),
-            label: Text("clear"),
-            action: () {
+            label: Text("clear", style: TextStyle(color: Colors.redAccent)),
+            onPull: () {
               _firstTextController.clear();
               _secondTextController.clear();
               _thirdTextController.clear();
@@ -104,7 +103,7 @@ class NextPageState extends State<NextPage> {
           ),
         ],
         child: ListView(
-          physics: BouncingScrollPhysics(),
+          physics: AlwaysScrollableScrollPhysics(),
           children: [
             TextField(
               controller: _firstTextController,
@@ -115,7 +114,6 @@ class NextPageState extends State<NextPage> {
             TextField(
               controller: _thirdTextController,
             ),
-            ...List.generate(100, (index) => Text(index.toString(), style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold))),
           ],
         ),
       ),